+++ +sbfsem/+io/OData.m
@@ -1,14 +1,37 @@
 classdef OData < handle
+    % ODATA
+    %
+    % Description:
+    %   Parent class for all OData query subclasses, handles misc
+    %   standalone OData queries
+    %
+    % Constructor:
+    %   obj = sbfsem.io.OData(source)
+    % Inputs:
+    %   source      Volume name or abbreviation
+    %
+    % Protected properties:
+    %   source      Full volume name (char)
+    %   baseURL     Service root (char)
+    %   webOpt      OData-specific weboptions
+    %
+    % History:
+    %   17Dec2017 - SSP
+    %   5Mar2018 - SSP - Updated for new JSON decoder
+    % ---------------------------------------------------------------------
 
 	properties (Access = protected)
 		source
 		baseURL
+        webOpt
 	end
 
 	methods
 		function obj = OData(source)
+            % ODATA  Constructor
 			obj.source = validateSource(source);
 			obj.baseURL = [getServerName(), '/', 'OData/'];
+            obj.webOpt = getODataOptions();
 		end
 	end
 
@@ -20,8 +43,9 @@ classdef OData < handle
 
             locationURL = getODataURL(ID, obj.source, 'location');
             importedData = readOData(locationURL);
+            value = cat(1, importedData.value{:});
             if ~isempty(importedData.value)
-                Locs = obj.processLocationData(importedData.value);
+                Locs = obj.processLocationData(value);
             else
                 Locs = [];
                 % This is important to track bc throws errors in VikingPlot
@@ -36,11 +60,12 @@ classdef OData < handle
             
             linkURL = getODataURL(ID, obj.source, 'link');
             importedData = readOData(linkURL);
+            value = cat(1, importedData.value{:});
             if ~isempty(importedData.value)
-                LocLinks = zeros(size(importedData.value, 1), 3);
+                LocLinks = zeros(size(value, 1), 3);
                 LocLinks(:, 1) = repmat(ID, [size(importedData, 1), 1]);
-                LocLinks(:, 2) = vertcat(importedData.value.A);
-                LocLinks(:, 3) = vertcat(importedData.value.B);
+                LocLinks(:, 2) = vertcat(value.A);
+                LocLinks(:, 3) = vertcat(value.B);
             else
                 LocLinks = [];
             end
@@ -48,30 +73,30 @@ classdef OData < handle
 	end
 
 	methods
-		function data = annotationsBySection(obj, sections, expandLinks)
-            if nargin < 3
-                expandLinks = false;
-            else
-                assert(islogical(expandLinks), 'Expand links is t/f');
-            end
+		function data = annotationsBySection(obj, sections)
+            % ANNOTATIONSBYSECTION
+            % Input:
+            %   sections        range of Z sections (vector)
+            % TODO: expand links t/f option
 			str = ['/Locations?$filter=Z le ' num2str(max(sections)),...
 				' and Z ge ', num2str(min(sections)),...
 				' and TypeCode eq 1',...
                 '&$select=ID,ParentID,X,Y,Z,Radius'];            
 
-			data = webread([getServiceRoot(obj.source), str], weboptions);
+			data = webread([getServiceRoot(obj.source), str], obj.webOpt);
 		end
 
-		function data = getLinkedIDs(obj, ID, vitread)
+		function data = getLinkedIDs(obj, ID)
             % GETLINKEDIDS  From ID, get other linked location IDs
             % Inputs:
             %   ID          The location ID
             %   vitread     Direction (t/f), empty for both
+            
             str = [getServiceRoot(obj.source),...
                 '/Locations(' num2str(ID) ')',...
                 '&$expand=LocationLinksA'];
             
-            data = webread(str, weboptions);
+            data = webread(str, obj.webOpt);
         end
 	end