+++ util/import/readOData.m
@@ -1,12 +1,20 @@
 function data = readOData(endpoint)
-	% READODATA  Webread with options
-	% Inputs:
-	%	endpoint 	sql query url
-	% Use:
+	% READODATA  
+    %
+    % Description:
+    %   Webread with OData-specific options
+    %
+    % Syntax:
 	%	data = readOData(getODataURL(127, 'i', 'location'));
+    %
+	% Inputs:
+	%	endpoint 	OData query url
+    % Outputs:
+    %   data        Information returned from webread()
 	%
-	% Note: for some reason this always times out the first time
-	% 	it is used in a matlab window. Just run it again...	
+    % History:
+    %   13Nov2017 - SSP
+    %   5Mar2018 - SSP - Added getODataOptions function
+    % ---------------------------------------------------------------------
 	
-    data = webread(endpoint,...
-        weboptionsOData);
+    data = webread(endpoint, getODataOptions());