+++ util/import/weboptionsOData.m
@@ -1,5 +1,0 @@
-function opts = weboptionsOData()
-	% WEBOPT  OData webopts
-
-	opts = weboptions('Timeout', 60,...
-		'ContentType', 'json');